+++ src/antidote/core/__init__.py
@@ -2,10 +2,10 @@ from .annotations import From, FromArg, Get, Provide
 from .container import Container, DependencyValue, Scope
 from .injection import Arg, DEPENDENCIES_TYPE, inject
 from .provider import Provider, StatelessProvider, does_not_freeze
-from .utils import LazyDependency, DependencyDebug
+from .utils import DependencyDebug
 from .wiring import Wiring, WithWiringMixin, wire
 
 __all__ = ['Provide', 'Get', 'From', 'FromArg',
            'Container', 'DependencyValue', 'Scope', 'inject', 'Arg',
            'DEPENDENCIES_TYPE', 'does_not_freeze', 'Provider', 'StatelessProvider',
-           'LazyDependency', 'DependencyDebug', 'wire', 'Wiring', 'WithWiringMixin']
+           'DependencyDebug', 'wire', 'Wiring', 'WithWiringMixin']