+++ tests/core/test_utils.py
@@ -1,30 +1,7 @@
-import pytest
-
-from antidote import world
-from antidote.core import LazyDependency, DependencyDebug, DependencyValue, Scope
+from antidote.core import DependencyDebug, DependencyValue, Scope
 
 
-def test_dependency():
-    with world.test.empty():
-        world.test.singleton('x', object())
-        d = LazyDependency('x', object)
-        assert d.unwrapped == 'x'
-        assert d.get() is world.get('x')
-
-    class A:
-        pass
-
-    with world.test.empty():
-        world.test.singleton('a', A())
-        world.test.singleton('x', object())
-
-        assert LazyDependency('a', A).get() is world.get('a')
-
-        with pytest.raises(TypeError):
-            LazyDependency('x', A).get()
-
-
-def test_dependency_instance():
+def test_dependency_value():
     ref = DependencyValue("test", scope=Scope.singleton())
     assert ref == DependencyValue("test", scope=Scope.singleton())
     assert ref != DependencyValue("test2", scope=Scope.singleton())